CMA-37 [refact] : import 경로 통일, 컴포넌트 병합 작업 및 login로직 props drilling 처리
