--- src/components/pages/login/find-id-pw-link/FindIdPwLink.tsx
@@ -1,12 +1,0 @@
-import Link from 'next/link';
-import * as style from '@/components/pages/login/find-id-pw-link/FindIdPwLink.style';
-
-/**
- * TODO : 아이디 비밀번호 href 추가 필요
- */
-export const FindIdPwLink = () => (
-  <style.Wrapper>
-    <Link href="">아이디 찾기</Link>
-    <Link href="">비밀번호 찾기</Link>
-  </style.Wrapper>
-);
