--- src/components/pages/login/find-id-pw-link/index.tsx
@@ -1,3 +1,0 @@
-import { FindIdPwLink } from './FindIdPwLink';
-
-export default FindIdPwLink;
