+++ src/components/pages/login/find-link-box/index.tsx
@@ -1,0 +1,3 @@
+import { FindLinkBox } from './FindLinkBox';
+
+export default FindLinkBox;
