--- src/components/pages/login/login-btn/index.tsx
@@ -1,3 +1,0 @@
-import { LoginBtn } from './LoginBtn';
-
-export default LoginBtn;
