+++ src/components/pages/login/login-input-box/LoginInputBox.style.tsx
@@ -1,7 +1,9 @@
 import styled, { css } from 'styled-components';
+import PasswordVisibleSvg from '@/assets/images/passwordVisible.svg';
+import PasswordInvisibleSvg from '@/assets/images/passwordInvisible.svg';
 
 export const Wrapper = styled.div`
-  margin-bottom: 58px;
+  height: 191px;
 
   position: relative;
 `;
@@ -18,6 +20,8 @@ export const Button = styled.button`
       justify-content: center;
       align-items: center;
 
+      position: relative;
+      top: 21px;
       left: 0;
       right: 0;
 
@@ -32,4 +36,26 @@ export const Button = styled.button`
       border-radius: 10px;
     `;
   }}
+`;
+
+// SVG
+export const NowPasswordVisibleIcon = styled(PasswordVisibleSvg)`
+  position: absolute;
+  top: 54.2px;
+  right: 25px;
+  justify-content: center;
+  align-items: center;
+
+  cursor: pointer;
+`;
+export const NowPasswordInvisibleIcon = styled(PasswordInvisibleSvg)`
+  position: absolute;
+  top: calc(
+    54px - (4.43px / 2)
+  ); // showPasswordSvg 와 HidePasswordSvg의 크기가 달라서 차이를 계산
+  right: 25px;
+  justify-content: center;
+  align-items: center;
+
+  cursor: pointer;
 `;
