+++ src/components/pages/login/login-input-box/index.tsx
@@ -1,0 +1,3 @@
+import { LoginInputBox } from './LoginInputBox';
+
+export default LoginInputBox;
