--- src/components/pages/login/login-inputs/index.tsx
@@ -1,3 +1,0 @@
-import { LoginInputs } from './LoginInputs';
-
-export default LoginInputs;
