--- src/components/pages/login/main-logo/MainLogo.style.tsx
@@ -1,21 +1,0 @@
-import styled from 'styled-components';
-import CommaIcon from '@/assets/images/commaIcon.svg';
-
-export const Wrapper = styled.div`
-  width: 100%;
-  height: 344px;
-
-  position: relative;
-`;
-
-// SVG
-export const MainIcon = styled(CommaIcon)`
-  width: 153.49px;
-  height: 124.99px;
-  margin: 0 auto;
-
-  position: relative;
-
-  display: block;
-  top: 134px;
-`;
