--- src/components/pages/login/main-logo/MainLogo.tsx
@@ -1,9 +1,0 @@
-import * as style from './MainLogo.style';
-
-const MainLogo = () => (
-  <style.Wrapper>
-    <style.MainIcon />
-  </style.Wrapper>
-);
-
-export default MainLogo;
