--- src/components/pages/login/main-logo/index.tsx
@@ -1,3 +1,0 @@
-import MainLogo from './MainLogo';
-
-export default MainLogo;
