--- src/components/pages/login/register-link/index.tsx
@@ -1,3 +1,0 @@
-import { RegisterLink } from './RegisterLink';
-
-export default RegisterLink;
