--- src/components/pages/login/sns-login-box/sns-login-text-div/SnsLoginTextDiv.style.tsx
@@ -1,38 +1,0 @@
-import styled, { css } from 'styled-components';
-
-export const Wrapper = styled.div`
-  ${({ theme }) => {
-    const { fonts, colors } = theme;
-    return css`
-      width: 100%;
-      margin: 0 auto;
-
-      position: relative;
-
-      display: flex;
-      align-items: center;
-      text-align: center;
-      overflow: hidden;
-
-      color: ${colors.grayscale.dark};
-      font-size: 12px;
-      font-family: ${fonts.family.noto};
-      font-weight: ${fonts.weight.medium};
-      line-height: 12px;
-
-      ::before,
-      ::after {
-        height: 1px;
-        margin: 0px 16px;
-
-        flex-grow: 1;
-
-        content: '';
-        font-size: 0px;
-        line-height: 0px;
-
-        background: ${colors.grayscale.gray300};
-      }
-    `;
-  }}
-`;
