--- src/components/pages/login/sns-login-box/sns-login-text-div/SnsLoginTextDiv.tsx
@@ -1,3 +1,0 @@
-import * as style from './SnsLoginTextDiv.style';
-
-export const SnsLoginTextDiv = () => <style.Wrapper>간편 로그인</style.Wrapper>;
