--- src/components/pages/login/sns-login-box/sns-login-text-div/index.tsx
@@ -1,3 +1,0 @@
-import { SnsLoginTextDiv } from './SnsLoginTextDiv';
-
-export default SnsLoginTextDiv;
