+++ src/components/template/login/LoginTemplate.style.tsx
@@ -1,8 +1,27 @@
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
+import CommaIcon from '@/assets/images/commaIcon.svg';
 
 export const Wrapper = styled.div`
   width: 100%;
   height: 100%;
 
   position: relative;
+`;
+
+export const Box = styled.div`
+  width: 100%;
+  height: 344px;
+
+  position: relative;
+`;
+
+// SVG
+export const MainIcon = styled(CommaIcon)`
+  width: 153.49px;
+  height: 124.99px;
+  margin: 134px auto 85.01px;
+
+  position: relative;
+
+  display: block;
 `;
